requirements.txt
crewai
crewai-tools
python-dotenv
streamlit
chromadb
sqlite3>=3.35.0